EGATE-89: Separated the Activiti and ACS resource POJOs into their own packages.
