+++ src/main/java/org/alfresco/event/model/activiti/TaskResourceV1.java
@@ -13,10 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.alfresco.event.model;
+package org.alfresco.event.model.activiti;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
+
+import org.alfresco.event.model.HierarchyEntry;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -25,55 +28,36 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
  * @author Jamal Kaabi-Mofrad
  */
 @JsonInclude(Include.NON_NULL)
-public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
+public class TaskResourceV1 extends ActivitiCloudRuntimeResourceV1
 {
-    // Only relevant for certain events e.g. the filed won't be
-    // included in the Activity Completed event.
-    private String activityName;
-    private String activityType;
-    private String elementId;
+    private String name;
     private String processDefinitionId;
     private String processInstanceId;
-    // Only relevant for Activity Cancelled event.
+    private Integer priority;
+    private String status;
+    private String assignee;
+    private Date createdDate;
+    private Date claimedDate;
+    // Only relevant for Task Cancelled event.
     private String cause;
 
-    public ActivityResourceV1()
+    public TaskResourceV1()
     {
     }
 
-    public ActivityResourceV1(String id, List<HierarchyEntry> primaryHierarchy)
+    public TaskResourceV1(String id, List<HierarchyEntry> primaryHierarchy)
     {
-        super(id, ActivityResourceV1.class, primaryHierarchy);
+        super(id, TaskResourceV1.class, primaryHierarchy);
     }
 
-    public String getActivityName()
+    public String getName()
     {
-        return activityName;
+        return name;
     }
 
-    public void setActivityName(String activityName)
+    public void setName(String name)
     {
-        this.activityName = activityName;
-    }
-
-    public String getActivityType()
-    {
-        return activityType;
-    }
-
-    public void setActivityType(String activityType)
-    {
-        this.activityType = activityType;
-    }
-
-    public String getElementId()
-    {
-        return elementId;
-    }
-
-    public void setElementId(String elementId)
-    {
-        this.elementId = elementId;
+        this.name = name;
     }
 
     public String getProcessDefinitionId()
@@ -96,6 +80,56 @@ public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
         this.processInstanceId = processInstanceId;
     }
 
+    public Integer getPriority()
+    {
+        return priority;
+    }
+
+    public void setPriority(Integer priority)
+    {
+        this.priority = priority;
+    }
+
+    public String getStatus()
+    {
+        return status;
+    }
+
+    public void setStatus(String status)
+    {
+        this.status = status;
+    }
+
+    public String getAssignee()
+    {
+        return assignee;
+    }
+
+    public void setAssignee(String assignee)
+    {
+        this.assignee = assignee;
+    }
+
+    public Date getCreatedDate()
+    {
+        return createdDate;
+    }
+
+    public void setCreatedDate(Date createdDate)
+    {
+        this.createdDate = createdDate;
+    }
+
+    public Date getClaimedDate()
+    {
+        return claimedDate;
+    }
+
+    public void setClaimedDate(Date claimedDate)
+    {
+        this.claimedDate = claimedDate;
+    }
+
     public String getCause()
     {
         return cause;
@@ -109,8 +143,8 @@ public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
     @Override
     public int hashCode()
     {
-        return Objects.hash(super.hashCode(), activityName, activityType, elementId,
-                    processDefinitionId, processInstanceId, cause);
+        return Objects.hash(super.hashCode(), name, processDefinitionId, processInstanceId,
+                    priority, status, assignee, createdDate, claimedDate, cause);
     }
 
     @Override
@@ -120,7 +154,7 @@ public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
         {
             return true;
         }
-        if (!(obj instanceof ActivityResourceV1))
+        if (!(obj instanceof TaskResourceV1))
         {
             return false;
         }
@@ -128,12 +162,15 @@ public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
         {
             return false;
         }
-        ActivityResourceV1 that = (ActivityResourceV1) obj;
-        return Objects.equals(activityName, that.activityName)
-                    && Objects.equals(activityType, that.activityType)
-                    && Objects.equals(elementId, that.elementId)
+        TaskResourceV1 that = (TaskResourceV1) obj;
+        return Objects.equals(name, that.name)
                     && Objects.equals(processDefinitionId, that.processDefinitionId)
                     && Objects.equals(processInstanceId, that.processInstanceId)
+                    && Objects.equals(priority, that.priority)
+                    && Objects.equals(status, that.status)
+                    && Objects.equals(assignee, that.assignee)
+                    && Objects.equals(createdDate, that.createdDate)
+                    && Objects.equals(claimedDate, that.claimedDate)
                     && Objects.equals(cause, that.cause);
     }
 
@@ -153,11 +190,14 @@ public class ActivityResourceV1 extends ActivitiCloudRuntimeResourceV1
                     .append(", \"serviceType\": ").append("\"" + this.serviceType + "\"")
                     .append(", \"entityId\": ").append("\"" + this.entityId + "\"")
                     .append(", \"timestamp\": ").append("\"" + this.timestamp + "\"")
+                    .append(", \"name\": ").append("\"" + this.name + "\"")
                     .append(", \"processDefinitionId\": ").append("\"" + this.processDefinitionId + "\"")
                     .append(", \"processInstanceId\": ").append("\"" + this.processInstanceId + "\"")
-                    .append(", \"activityName\": ").append("\"" + this.activityName + "\"")
-                    .append(", \"activityType\": ").append("\"" + this.activityType + "\"")
-                    .append(", \"elementId\": ").append("\"" + this.elementId + "\"")
+                    .append(", \"priority\": ").append(this.priority)
+                    .append(", \"status\": ").append("\"" + this.status + "\"")
+                    .append(", \"assignee\": ").append("\"" + this.assignee + "\"")
+                    .append(", \"createdDate\": ").append("\"" + this.createdDate + "\"")
+                    .append(", \"claimedDate\": ").append("\"" + this.claimedDate + "\"")
                     .append(", \"cause\": ").append("\"" + this.cause + "\"")
                     .append("}}");
         return builder.toString();
